Update Vuetify, AppBar/Footer. Implement ModalScreen shell. Fix bugs.
ModalScreen is the shell used to navigate Product, WorkOrder and other
entities details. Hides app bar and footer to provide focus, and presents
navigation links on the top right.

The component is implemented as an empty shell used in a specific base
route which is useless by itself, but has its own children routes which
represents the different links on the top right.

Still need to test content overflow and implement automated link
injection.