Implement single piece tracking and WIP creation at `batch_complete`
Considering that each job generates its own batches, the serial no.
will belong to multiple batches as it goes through the process.

So the serial record contains a list of the batches it belongs to.

Added also to the batch record the list of serials that were processed
within it.

WIP records are at the batch level and contain a list of the serial
numbers included.