Close WO only when all jobs have been closed
Initially WO would be closed when the completed quantity from the last
phase equalled the WO quantity. This created an issue in case there
would be jobs in previous phases still to be completed.

The above condition theoretically should not happen considering that
the last phase should build upon the results of the previous one,
but in this way the process is more robust while allowing for greater
flexibility.