Use url rather than subdomain for service routing
It's possible that in production an IP address will be used instead of
a domain name, so subdomains might not be feasible. Url path is a more
robust solution that can always work.

Root path in the entry file for the api server must be provided for
correct retrieval of the api docs.