Implement manual new batch and auto_new_batch option (#185)
This allows for simpler thinking. Now work sessions and batches cannot
overlap. A batch may have multiple work sessions if the user pauses
the job with an active batch ongoing, but a work session will always
refer to one and one batch only.

When a batch is completed, the work session is closed. If the
auto_new_batch option is on, then a new batch and a related work session
are created. If not, the job is paused and references to the current
batch in it is removed.

This required breaking changes in the event class and traceability
queries, linked to changes in the models and the general logic.

some utility functions returned data from the db, letting the calling
method deal with saving data as event properties when necessary.

However this required duplicating code in multiple methods of the class,
e.g. saving the work session key in the event `info` property.

The refactoring moved the saving of this key data into event properties
directly within the utility functions, so now in methods is sufficient
to call e.g. `self.get_job_data()`.