Close #185 and close #189
-> 185: Close job when remaining quantity is changed to zero and there
is zero active quantity (no active batch).
Added tracking of active batch quantity within the job
Changed approach: the job was previously flagged for deletion. Now it's
just flagged as closed with no more remaining quantity and a note
specifying who closed the job when.

-> 189: Make automatic start of new batch optional
The alternative is that the job is paused at the completion of a batch.
In this case the new batch is created when the job is resumed.

Requires adding a parameter to phases and jobs.

Added `qt_total` in Batch model, corresponding to the
`active_batch_qt` in the job

!!! BREAKING CHANGES !!!
- The BatchTimeRecord collection has been removed in favor of a simpler
approach using only WorkSessions which cannot span multiple batches
anymore. In case of automatic start of a new batch, the current
WorkSession is closed and a new one is created.

- The attribute `current_batch` has been substituted by a more
expressive `active_batch_key`

BONUS:
- remove unnecessary print statements
- add reset of production and traceability data as an admin endpoint
- refactor `update_target_queue` into utils module for better reuse
- refactor Event methods to store updates in the object rather than
  returning results. Improves DRY approach
- improve responses and exception/transaction handling
  * `get_assingment_list`
  * `update_jobs`
  * `get_batch_execution_data`
  * `verify_user`
  * `save`, `complete_step` methods of `Event` class
- removed unnecessary WorkSession management in the client (from
  session.js Vuex module)