Fix production_batch/step_check management. Closes #285
The step_check parameter is now a boolean. Batch quantity is solely
handled by the production_batch_qt. If this is set to zero it means
batch quantity is equal to job quantity. This allows for single
declaration for job also when no step check is not enabled. Closes #243

Step check could not be activated when no step is present in the
procedure. But if steps were deleted, step_check could remain active.
Modified the process update procedure so that step_check is forced
as false when no steps are present. Closes #273

Also fixed wip management (next_batch_qt definition) for single
declaration jobs (production_batch_qt = job quantity). Closes #284