Implement basic Issue Management
Issues can be created by operators in the job screen and managed both in
work session screen and in the work order screen.

Issues can be created without type, which can be assigned later. No
metadata needs to be provided at creation for now. Later issue types
will include optional forms to fill in with fields that can be marked
as required.

Includes also:
- Refactoring of event management code
- Improvement of exit procedure from WorkOrderScreen
- Better date formatting management
- Navigation drawer now can be used from modal screens too
- Disable annotations layer in PDFs