Allow override of job traceability data: duration/progress/active_batch
Logic:
- Only closed jobs can have their duration overridden
- Only assigned jobs can have progress set
- Progress updates must respect upstream and downsrteam wip available
  and can be done only if there is no active batch (if there is, the
  active batch must/can be canceled first)

Overridden Batches and WorkSessions are flagged as `canceled` and
manually created ones are flagged as `forced`. Both `canceled` and
`forced` fields store the event id that triggered the override. Jobs
subject to an override are also flagged as forced with the event id.

Overridden wip is deleted or updated.

Progress overrides take care of job and work order status updates
(from created to started to closed and viceversa) as well as adding or
removing jobs and the work order from queues.

Canceled work sessions and batches do not account for performance
measurements, but are kept for auditing purposes and other statistics
potentially interesting such as frequency of errors/overrides.

TO DO: Handle work order start date management. The end date is taken
care of by the UPDATE_WORK_ORDER query, but the start must be handled
separately.

Includes various fixes:
- Add endpoint to get wip availability for job
- restore dynamic calculation of processing time for work orders to
  update it in real time in the UI
- Fix ProgressBtn/WorkSession behavior in managing heartbeat and job
  closing/exit procedure
- Use int for work session duration model for easier calculations
- Removed redundant std_processing_time property from phase_model
  (already present in params)
- List only active products in WorkOrderNew
- Add NoDataAlert to JobList
- Prevent unnecessary wip alert when declaring last batch
- Improve layout in WorkOrderDataColumn
- ensure available wip is retrieved for specific work order
- Refactored event code architecture
- Improve error handling in events