Implement form field admin feature + general admin improvements
BACKEND
- added field edit/delete endpoints
- added create/edit/delete endpoints for list field values
- added document _key/_id fields to CustomListValue model

Notes: the changes in the DB queries are necessary to avoid errors in
existing field instances of which the definition has been deleted.
This way old data can be silently persisted in the database without
showing the field anymore and field deletion doesn't require complex
backend validation or waterfall updates.

FRONTEND
- Added field admin section to handle both field metadata and list
  values
- Improved other admin sections to avoid race conditions and missing
  object props while rendering the first item selected
- Aligned styling of search fields to filled
- Renamed CustomFieldNew component to FormFieldNew to align with other
  components
- Improved BaseActionCard parametrizing button labels and colors

Closes #309, #342