feat: include print templates in IssueType model
Required creating a new model for issue type updates

Issue form field model required update because it was not enforced
and was saved using the _key attribute instead of field_key due to a
mistake in the frontend. Could have modified the frontend, but there
are existing customers with data using the wrong model so to avoid
a db update I stuck to using the _key attribute and updated the model
accordingly. Also added the type attribute that was missing.