fix: wip management for custom batch declaration
the booking of additional wip for declarations higher than the active
batch must be done before the creation of a new batch, otherwise wip
is going to be booked twice.

In create_batch method, job data must be fetched only if no already
present otherwise it will overwrite updates that have not yet been
persisted in the database, such as new completed quantity or new
active batch.