fix: remove retrieval of booked but inactive wip
a filter was commented out for serials management, which caused a
regression considering inactive wip for the work order in ANY phase as
available for the jobs indicated, making next_batch_available = True
even if it isn't.

Instead of fixing it, the part is removed because currently the
anticipated booking of product wip (booked but not active) is not
possible, and not sure will ever happen (may happen in the future for
bom components).