fix: navigation to ProductScreen from non-library routes
the system saved the navigation state for each product in the product
list. Even if the specific product was added to the list, the
ProductScreen was configured as a child of the ProductList view which
then fired the standard fetching of the product library, which would
overwrite the list.

The solution thus implied:
- appending the product with the nav state in the list if necessary
- also removing the product screen as a child of the ProductList route.
- the data sources in ProductionProcess

Ref: closes #557