wip: refactor print context classes and prep for image management
placed all the switch logic in the base `TemplateContext` class and
started updating constructors to ensure relevant data is present.

started trying to show images in the PrintDialog

ideally this should lead to move the work session prints from the step
to the batch, being able to fetch data from all the steps at once
and print from anywhere in the procedure, possibly adding the chance to
force printing before declaring.

To do so the template assignment to should be moved from the step to the
phase and we should add new parameter for mandatory printing and a job
attribute to keep track of it.

Prints could be in a new tab of the WorkSession