fix: transfer inventory update
update existing record instead of creating new one

Previously inventory records would be aggregated by the app, but not
anymore. It makes sense to update them, and since we're updating it,
remove the date received and expiration date, they can't be tracked and
are not used anyway at the moment. In the future they will require a
lot/serial number.