fix: avoid bom data reset when starting job
job data is taken from the wo_bom property of the job which is computed
via query and was not done when responding to the JOB_STARTED event,
resetting temporarily the bom until the next refresh.

This caused serious issues in case the bom required traceability data
for components, allowing operators to declare batches without needed info.