fix: prevent empty file metadata in FilesList
in some cases, e.g. in case of page refresh, file metadata is kept,
showing the filename listed in the file FormField component while
the file was not actually available, which may mislead the user to think
the file would be saved.

This fix deletes file metadata if the content is not available.