fix: serial consumption movements
when component serials are linked the movement is already generated,
so trying to create it again in BatchCompleted raises errors (serial
not available in inventory anymore).

temporary solution is to not process serial consumption movements in
BatchCompleted.

TODO: it would be better to generate it in the BatchCompleted, for
consistency and more streamlined integrations, but to do so we need a
way to distinguish "finalized" link/unlink, from temporary
ones carried out in work session screen which is more complicated.

We could also simply generate two events in the client when we want the
link/unlink to also generate the movement, but that would be another
workaround, it makes sense that the link/unlink triggers the movement
directly, or at least can in some contexts.