fix: new batch creation with traceability
completed batch serials were being passed down to the BatchCreated
event via the self.info dump in the `create_as_child` method, because
both have a `batch_serial` property.

This led to new batch serials being created with a code equal to the
key of the previous batch serial.

Renamed the BatchCreated property to avoid conflict.