fix!: serial data management
previous setup created issues. Serials in work orders created before
the latest process update at the product level would find inconsistencies
because the update algorithm looked up the product process, while
step form fields stored in the job referred to the old ones before the
update.

This led to serial data not being updated correctly because step
data would not find a match in the current product process.

In general it was necessary to store serial fields data in the work
order so that process changes do not impact the serial creation of
that work order.

TODO: implement product versioning in order to avoid this issue and
associate the serial to a specific version.

UPGRADE: requires:

- rebuilding serial data to add metadata to each field and if necessary
  add the value from the work order execution data.

- adding serial data template to all open work orders, based on their
  job form fields

See issue for more details.

ref: #684