fix: unauthorized routes handling
the previous management made it so that a if user navigated to an
unauthorized route directly (via url) it would not get out of it.

this way navigation will be interrupted if it's coming from a link in
another page, and otherwise will be redirected to their homepage.

also added scope to the task screen route